Add files via upload
A machine that can perform a complicated series of tasks automatically “and “A machine that is made to look like a human and that can do some things that a human can do” – These are the defining terms found in the entry against the word ‘ROBOT’ in the famous ‘Oxford Advanced Learner’s Dictionary of current English’ by Astheny. Robot is a system containing sensors, control systems, manipulations, power suppliers and software – all working together like an organized team effort to perform a task. Fire-fighting robots are useful in situations where it is too difficult to extinguish fire. It is also be very useful in situations where fire fighters have difficulty in reaching the place which is of fire and to be extinguished. Modular design based on proximity, vision and IR sensors has been incorporated into a fire-fighting model. The robot has a sprinkler in it and it is useful in extinguishing flame by sprinkling water in the place of fire. Software consists of integrated tracking, obstacle avoidance, flame detection, motion algorithms and detection. Through testing, it is possible to run in a realistic scene simulated in the lab and to detect and extinguishment the flame. Robot can transfer video to remote location.